Cleaning Trace Shader (Hindered Progress)
Managed to "fix" the shader and how it responds to moving roomba. For now there are placeholder brick and wood texture for floors to test the functionality between different maps. TextureSetter.cs script is the one that fetches the live texture update from the [SYSTEM] Ground Texture Generator.

Sad post-mortem: After pulling 13 changes from github, the little testee roomba doesn't want to move and shader is not displayed on the floor. FML.

Next up:
1. apply appropriate textures for floors.
2. Import the floor into the main game

Important note about the code:
TextureSetter.cs is still janky and doesn't properly fetch the Texture2D "MaskCleaner" within the "Shader Graph Floor". It doesn't go into the conditionals even though trying to get the texture outputs null. It is broken, but doesn't produce errors (??). Something to watch out in case shader is broken again.

Note about the shader:
There was a need to rotate the mask map for 180 degrees to follow the roomba. Otherwise it would go in opposite direction. This is a quick bug fix, and in case shader doesn't respond as intended look into the shader to remove the UV rotation node.